--- src/tags/TagManager.js
@@ -1,14 +1,0 @@
-import { TagForm } from "./TagForm"
-import { TagList } from "./TagList"
-
-
-export const TagManager = () => {
-
-
-    return <article>
-        <TagList />
-        <TagForm />
-
-    </article>
-    
-}
